analyser/requirements.txt
watchdog
requests
fastapi
uvicorn
pydantic
python-dotenv
psycopg2-binary
